# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "F1 Teams API",
    "description": "API for managing F1 teams, drivers, and their details.",
    "version": "1.0.0",
    "contact": {
      "email": "fran.sribar@fer.hr"
    },
    "license": {
      "name": "Creative Commons Zero v1.0 Universal"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "Serve the home page",
        "responses": {
          "200": {
            "description": "HTML content of the home page"
          }
        }
      }
    },
    "/datatable": {
      "get": {
        "summary": "Serve the datatable page",
        "responses": {
          "200": {
            "description": "HTML content of the datatable page"
          }
        }
      }
    },
    "/add-team": {
      "get": {
        "summary": "Serve the add team page",
        "responses": {
          "200": {
            "description": "HTML content of the add team page"
          }
        }
      }
    },
    "/delete-team": {
      "get": {
        "summary": "Serve the delete team page",
        "responses": {
          "200": {
            "description": "HTML content of the delete team page"
          }
        }
      }
    },
    "/update-team/{teamId}.html": {
      "get": {
        "summary": "Serve the update team page",
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "HTML content of the update team page"
          },
          "404": {
            "description": "Team update page not found"
          }
        }
      }
    },
    "/drivers": {
      "get": {
        "summary": "Serve the drivers page",
        "responses": {
          "200": {
            "description": "HTML content of the drivers page"
          }
        }
      }
    },
    "/details/{teamId}.html": {
      "get": {
        "summary": "Serve the team details page",
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "HTML content of the details page"
          },
          "404": {
            "description": "Details page for the team not found"
          }
        }
      }
    },
    "/api/F1teams": {
      "get": {
        "summary": "Fetch filtered F1 teams and drivers",
        "parameters": [
          {
            "name": "filter_value",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter_field",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of F1 teams and drivers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No matching F1 teams or drivers found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/download/csv": {
      "get": {
        "summary": "Download filtered data as CSV",
        "parameters": [
          {
            "name": "filter_value",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter_field",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CSV file generated successfully",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "No data available to export as CSV"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/download/json": {
      "get": {
        "summary": "Download filtered data as JSON",
        "parameters": [
          {
            "name": "filter_value",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter_field",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JSON file generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No data available to export as JSON"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/addNewTeam": {
      "post": {
        "summary": "Add a new F1 team and drivers",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "teamName": {
                    "type": "string"
                  },
                  "engine": {
                    "type": "string"
                  },
                  "licensedIn": {
                    "type": "string"
                  },
                  "seasonEntered": {
                    "type": "integer"
                  },
                  "driver1Name": {
                    "type": "string"
                  },
                  "driver1Surname": {
                    "type": "string"
                  },
                  "driver1BirthYear": {
                    "type": "integer"
                  },
                  "driver1Nationality": {
                    "type": "string"
                  },
                  "driver2Name": {
                    "type": "string"
                  },
                  "driver2Surname": {
                    "type": "string"
                  },
                  "driver2BirthYear": {
                    "type": "integer"
                  },
                  "driver2Nationality": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Team and drivers added successfully"
          },
          "400": {
            "description": "Invalid request body"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/getAllTeams": {
      "get": {
        "summary": "Fetch all teams",
        "responses": {
          "200": {
            "description": "A list of all teams",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No teams found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/deleteTeam/{teamId}": {
      "delete": {
        "summary": "Delete a team and its drivers",
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Team and drivers deleted successfully"
          },
          "404": {
            "description": "Team not found or already deleted"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/updateTeam/{teamId}": {
      "put": {
        "summary": "Update a team's details",
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Team updated successfully"
          },
          "400": {
            "description": "No valid fields provided to update"
          },
          "404": {
            "description": "Team not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/drivers": {
      "get": {
        "summary": "Fetch all drivers",
        "responses": {
          "200": {
            "description": "A list of all drivers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No drivers found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/details/{teamId}": {
      "get": {
        "summary": "Fetch details of a specific team",
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Details of a specific team",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Team not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  }
}